// src/constants/enums/searchBar.ts
export enum SearchBarForm {
  id = "search_bar__from_id",

  autocomplete_inputs = "off",
  input_size = "medium",
  button_size = "large",

  search_bar_input_id = "search_bar_input_id",
  search_bar_input_name = "search_bar_input",
  search_bar_input_label = "",
  search_bar_input_role = "search_bar-input",
  search_bar_input_arial_label = "search_bar-input",
  search_bar_input_placeholder = "Search in nikafeis",
  search_bar_input_type = "text",
}